e2e: spell the lifecycle steps as the product spells its verbs
The step was "run", which the product already reserves for a different
meaning - apply then start - while its actual verb is "start". The envelope
mapped run onto START and rejected start outright, so the one word an author
would reach for first was an error, and the frozen surface disagreed with the
product about what "run" means. Spell the four steps exactly as the product
spells them; the hand-written switch that existed only to perform the rename
is gone, and "run" stays free for the meaning the product already gives it.

The state matcher is now a scalar. A specification references exactly one
pipeline and the executor already resolves its id through the product's
parser, so the map key was an id to copy by hand, get wrong, and then watch
poll a pipeline that does not exist until it timed out.

Close the gaps that let a specification pass while testing less than it says:

- A duplicate key last-won with only a log line, so a repeated table quietly
  dropped the first count and a repeated matcher word defeated the
  one-word-per-matcher guard. Duplicate keys are now rejected.
- count and state maps went through Map.copyOf, whose iteration order is
  randomised per JVM run, so endpoints were read from a live database in a
  different order every run and a failure message never read the same twice.
- An empty count body, and an absent or empty steps list, held vacuously.
- A fractional row count truncated in silence; a negative one was accepted.
- A key YAML resolved to a non-string escaped as a raw ClassCastException,
  naming nothing.
- An alias required exactly one dot, so the harness could not address a table
  whose name contains one - a table the product itself resolves by splitting
  at the first dot.
- A cdc row count was an int while a seed row count was a long.
- Wrapping every RuntimeException from the product's parser blamed the author
  for what may be a defect in the parser; only a DSL error is an authoring
  error now.
- A timed-out wait reported its configured bound rather than how long it
  actually waited, in the one place that sells honesty.

Tests: the immutability test asserted List.copyOf(x).equals(x), true of every
list by contract and unable to fail. The timeout test asserted the expected
and actual counts separately and so could not tell them apart when swapped.
The state spy discarded the id it was asked for. Each now pins the thing it
is named for.

+++ e2e/src/test/java/io/cyntex/e2e/Step.java
@@ -5,10 +5,13 @@ import io.cyntex.core.lifecycle.LifecycleVerb;
 /**
  * One stage of a specification. Steps run in declaration order; the order is the scenario.
  *
- * <p>Lifecycle steps carry the product's own verb enum rather than a private copy, so the step
- * vocabulary cannot drift from the verbs the product actually accepts. There is no rewind step:
- * re-snapshotting is the explicit {@code stop} then {@code run} pair, which is exactly what the
- * product's verb set offers.
+ * <p>Lifecycle steps are spelled exactly as the product spells them - {@code start}, {@code pause},
+ * {@code resume}, {@code stop} - and carry the product's own verb enum, so neither the word nor the
+ * value can drift from what the product accepts. {@code run} is deliberately not a step: the
+ * product already reserves it for a different meaning, apply-then-start.
+ *
+ * <p>There is no rewind step: re-snapshotting is the explicit {@code stop} then {@code start} pair,
+ * which is exactly what the product's verb set offers.
  */
 public sealed interface Step {
 
@@ -16,7 +19,7 @@ public sealed interface Step {
     record Lifecycle(LifecycleVerb verb) implements Step {}
 
     /** Produces changes against a seeded table while the pipeline is running. */
-    record Cdc(TableAlias table, CdcOp op, int rows) implements Step {}
+    record Cdc(TableAlias table, CdcOp op, long rows) implements Step {}
 
     /** Polls a matcher until it holds or the bound expires. */
     record Await(Matcher matcher) implements Step {}
